% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{class_size}
\alias{class_size}
\title{Class size data}
\format{
A tibble with 8064 rows/studies and 27 variables/columns
}
\usage{
class_size
}
\description{
Data from a systematic review on the effects of class size on student achievement in elementary school
(Filges, Sonne-Schmidt, & Nielsen, 2018).
}
\references{
Filges, T., Sonne-Schmidt, C. S., & Nielsen, B. C. V. (2018).
Small class sizes for improving student achievement in primary and secondary schools:
A systematic review.
\emph{Campbell Systematic Reviews}, 14(1), 1-107. \doi{10.4073/csr.2018.10}
}
\keyword{datasets}
