% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs.R
\name{df_puste}
\alias{df_puste}
\title{'Degrees of freedom calculation for cluster bias correction with Pustejovsky formula'}
\usage{
df_puste(N_total, ICC, avg_grp_size = NULL, n_clusters = NULL)
}
\arguments{
\item{N_total}{Numerical value indicating the total sample size of the study.}

\item{ICC}{Numerical value indicating the intra-class correlation (ICC) value.}

\item{avg_grp_size}{Numerical value indicating the average cluster size/
the average number of individuals per cluster.}

\item{n_clusters}{Numerical value indicating the number of clusters.}
}
\value{
Returns a numerical values indicating the cluster adjusted degrees of freedom.
}
\description{
This function calculates the degrees of freedom for studies
with clustering, using the upsilon formula from Pustejovsky (2016). Find under
the Cluster randomized trials section.
}
\examples{
df_puste(N_total = 100, ICC = 0.1, avg_grp_size = 5)
}
\references{
Pustejovsky (2016).
Alternative formulas for the standardized mean difference.
\url{https://www.jepusto.com/alternative-formulas-for-the-smd/}
}
