% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Dietrichson2021_data}
\alias{Dietrichson2021_data}
\title{Targeted school-based interventions data (K-6)}
\format{
A tibble with 1334 rows/studies and 71 variables/columns

\tabular{lll}{
\bold{Authors} \tab \code{character} \tab variable with short-hand information about authors. \cr
\bold{Study_ID} \tab \code{integer} \tab Number identifying the study cluster. \cr
\bold{Title}  \tab \code{character}   \tab Title of report/journal article. \cr
\bold{Outlet} \tab \code{character}   \tab Variable with information about outlet for the study, i.e., either a journal or a dissertation or a report etc. \cr
\bold{Country}    \tab \code{character}   \tab Variable with information about the country where the intervention was conducted. \cr
\bold{Language}  \tab \code{character}   \tab Variable with information about the language used in the study. \cr
\bold{Publishing_status}  \tab \code{character}   \tab Indicator equal to one if the study has been published in a journal. \cr
\bold{RCT_QES}  \tab \code{character}   \tab Variable denoting whether the effect size is from a RCT or a QES \cr
\bold{Level_treatment_assignment}  \tab \code{character}   \tab Variable indicating the level of treatment assignment. \cr
\bold{Test_subject}  \tab \code{character}   \tab Variable denoting the subject of test the effect size is based on (reading or math). \cr
\bold{Estimation_method}  \tab \code{character}   \tab Variable indicating the type of estimation method the effect
estimate is taken from (e.g., raw means, adjusted means, regression adjusted). \cr
\bold{Effectsize_g}  \tab \code{numeric}   \tab Effect size, Hedges' g, not adjusted for clustering. \cr
\bold{SE_g}  \tab \code{numeric}   \tab The standard error of the effect size, not adjusted for clustering. \cr
}
}
\usage{
Dietrichson2021_data
}
\description{
Data from a meta-analysis of the effects of targeted school-based interventions
on reading and mathematics for students with or at risk of academic difficulties in Grades K-6
(Dietrichson et al., 2021)
}
\references{
Dietrichson, J., Filges, T., Seerup, J. K., Klokker, R. H.,
Viinholt, B. C. A., Bøg, M., & Eiberg, M. (2021).
Targeted school-based interventions for improving reading and mathematics for
students with or at risk of academic difficulties in Grades K-6: A systematic review.
\emph{Campbell Systematic Reviews}, 17(2), e1152. \doi{10.1002/cl2.1152}
}
\keyword{datasets}
