% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_calc.R
\name{OR_calc}
\alias{OR_calc}
\title{Calculate odds ratios (OR)}
\usage{
OR_calc(
  A = NULL,
  B = NULL,
  C = NULL,
  D = NULL,
  p1 = NULL,
  p2 = NULL,
  n1 = NULL,
  n2 = NULL,
  OR = NULL,
  LL_OR = NULL,
  UL_OR = NULL,
  SE_OR = NULL,
  V_OR = NULL,
  Z = 1.96,
  add_name_to_vars = NULL,
  vars = dplyr::everything()
)
}
\arguments{
\item{A}{Insert}

\item{B}{Insert}

\item{C}{Insert}

\item{D}{Insert}

\item{p1}{Insert}

\item{p2}{Insert}

\item{n1}{Insert}

\item{n2}{Insert}

\item{OR}{Insert}

\item{LL_OR}{Insert}

\item{UL_OR}{Insert}

\item{SE_OR}{Insert}

\item{V_OR}{Insert}

\item{Z}{Insert}

\item{add_name_to_vars}{Optional character string to be added to the variables names of the generated \code{tibble}.}

\item{vars}{Variables to be reported. Default is \code{NULL}. See Value section for further details.}
}
\value{
A \code{tibble} with information about OR, OR_LN, vln_OR.
}
\description{
This function calculated odds ratios based on various type of input/information,
as described in Table 11.10 from Borenstein and Hedges (2019).
}
\details{
Add details\tabular{lllll}{
    \tab  \tab Event \tab Non-event \tab Total \cr
   Treated \tab  \tab A \tab B \tab n1 \cr
   Control \tab  \tab C \tab D \tab n2 \cr
}
}
\note{
Read Borenstein and Hedges (2019) for further details.
}
\examples{

# Using raw events
OR_calc(A = 20, B = 80, C = 10, D = 90)

# Using proportions
OR_calc(p1 = .53, p2 = .11, n1 = 20, n2 = 26)

# Using raw OR and CI
OR_calc(OR = 2.25, LL_OR = 1.5, UL_OR = 3)

# Adding suffix to variables and selecting specific variables
OR_calc(A = 20, B = 80, C = 10, D = 90, add_name_to_vars = "_test", vars = OR_test)


}
\references{
Borenstein and Hedges (2019).
Effect sizes for meta-analysis. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.),
\emph{The handbook of research synthesis and meta-analysis} (3rd ed., pp. 207–242).
Russell Sage Foundation West Sussex.
}
