% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgt_smd_1armcluster.R
\name{vgt_smd_1armcluster}
\alias{vgt_smd_1armcluster}
\title{Variance calculation when there is clustering in one treatment group only}
\usage{
vgt_smd_1armcluster(
  N_cl_grp,
  N_ind_grp,
  avg_grp_size,
  ICC,
  g,
  model = c("Posttest", "ANCOVA", "DiD", "Reg_coef", "Std_reg_coef"),
  not_cluster_adj = TRUE,
  prepost_cor = NULL,
  F_val = NULL,
  t_val = NULL,
  SE = NULL,
  SD = NULL,
  SE_std = NULL,
  R2 = NULL,
  q = 1
)
}
\arguments{
\item{N_cl_grp}{Insert}

\item{N_ind_grp}{Insert}

\item{avg_grp_size}{Insert}

\item{ICC}{Insert}

\item{g}{Insert}

\item{model}{Insert}

\item{not_cluster_adj}{Insert}

\item{prepost_cor}{Insert}

\item{F_val}{Insert}

\item{t_val}{Insert}

\item{SE}{Insert}

\item{SD}{Insert}

\item{SE_std}{Insert}

\item{R2}{Insert}

\item{q}{Insert}
}
\value{
A \code{tibble} with info ...(Insert)...
}
\description{
Insert
}
\details{
Insert
}
\note{
Insert
}
\examples{
vgt_smd_1armcluster(
N_cl_grp = 60, N_ind_grp = 40, avg_grp_size = 10, ICC = 0.1, g = 0.2,
model = "Posttest", not_cluster_adj = TRUE
)

}
\references{
Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}

What Works Clearinghouse (2021).
Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1
\emph{Institute of Education Science}.
\url{https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf}
}
\seealso{
\code{\link{df_h_1armcluster}}, \code{\link{eta_1armcluster}},
\code{\link{gamma_1armcluster}}
}
