% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs.R
\name{df_h_1armcluster}
\alias{df_h_1armcluster}
\title{Degrees of freedom calculation for cluster bias correction when there is clustering in one treatment group only}
\usage{
df_h_1armcluster(N_total, ICC, N_grp, avg_grp_size = NULL, n_clusters = NULL)
}
\arguments{
\item{N_total}{Numerical value indicating the total sample size of the study.}

\item{ICC}{Numerical value indicating the intra-class correlation (ICC) value.}

\item{N_grp}{Numerical value indicating the sample size of the arm/group containing clustering.}

\item{avg_grp_size}{Numerical value indicating the average cluster size.}

\item{n_clusters}{Numerical value indicating the number of clusters in the treatment group.}
}
\value{
Returns a numerical value indicating the cluster adjusted degrees of freedom.
}
\description{
This function calculates the degrees of freedom for studies
with clustering in one treatment group only, using Equation (7) from Hedges & Citkowicz (2015).
}
\details{
When clustering is present the \eqn{N-2}{N-2} degrees of freedom (\eqn{df}{df}) will be a rather liberal choice,
partly overestimating the small sample corrector \eqn{J}{J} and partly underestimating
the true variance of (Hedges') \eqn{g_T}{g-T}. The impact of the calculated \eqn{df}{df} will be most
consequential for small (sample) studies. To overcome these issues,
Hedges & Citkowicz (2015) suggest obtaining the degrees of freedom from

\deqn{ h = \frac{[(N-2)(1-\rho) + (N^T-n)\rho]^2}
{(N-2)(1-\rho)^2 + (N^T-n)n\rho^2 + 2(N^T-n)(1-\rho)\rho}}

where \eqn{N}{N} is the total sample size, \eqn{N^T}{N-T} is the sample size of the treatment group,
containg clustering, \eqn{n}{n} is average cluster size and
\eqn{\rho}{\rho} is the (imputed) intraclass correlation.
}
\note{
Read Taylor et al. (2020) to understand why we use the \eqn{g_T}{g-T} notation.
Find suggestions for how and which ICC values to impute when these are unknown (Hedges & Hedberg, 2007, 2013).
}
\examples{
df <- df_h_1armcluster(N_total = 100, ICC = 0.1, N_grp = 60, avg_grp_size = 5)
df


# Testing function
N <- 100
rho <- 0.1
NT <- 60
n <- 5

df_raw <- ((N-2)*(1-rho) + (NT-n)*rho)^2 /
          ( (N-2)*(1-rho)^2 + (NT-n)*n*rho^2 + 2*(NT-n)*(1-rho)*rho )

round(df_raw, 2)


}
\references{
Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}

Hedges, L. V., & Hedberg, E. C. (2007).
Intraclass correlation values for planning group-randomized trials in education.
\emph{Educational Evaluation and Policy Analysis}, 29(1), 60–87.
\doi{10.3102/0162373707299706}

Hedges, L. V., & Hedberg, E. C. (2013).
Intraclass Correlations and Covariate Outcome Correlations for Planning
Two- and Three-Level Cluster-Randomized Experiments in Education.
\emph{Evaluation Review}, 37(6), 445–489. \doi{10.1177/0193841X14529126}

Taylor, J.A., Pigott, T.D., & Williams, R. (2020)
Promoting Knowledge Accumulation About Intervention Effects:
Exploring Strategies for Standardizing Statistical Approaches and Effect Size Reporting.
\emph{Educational Researcher}, 51(1), 72-80. \doi{10.3102/0013189X211051319}
}
