% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_h_1armcluster.R
\name{df_h_1armcluster}
\alias{df_h_1armcluster}
\title{'Cluster bias correction when there is clustering in one treatment group only'}
\usage{
df_h_1armcluster(N_total, ICC, N_grp, avg_grp_size)
}
\arguments{
\item{N_total}{Numerical value indicating the total sample size of the study.}

\item{ICC}{Numerical value indicating the intra-class correlation (ICC) value.}

\item{N_grp}{Numerical value indicating the sample size of the arm/group containing clustering.}

\item{avg_grp_size}{Numerical value indicating the average cluster size.}
}
\value{
Returns a numerical values indicating the cluster adjusted degrees of freedom.
}
\description{
This function calculates the degrees of freedom for studies
with clustering in one treatment group only, using Equation (7) from Hedges & Citkowicz (2015).
}
\examples{
df_h_1armcluster(N_total = 100, ICC = 0.1, N_grp = 60, avg_grp_size = 5)
}
\references{
Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}
}
