% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_factors_1armcluster.R
\name{gamma_1armcluster}
\alias{gamma_1armcluster}
\title{Small number of clusters correction when there is clustering in one treatment group only}
\usage{
gamma_1armcluster(N_total, Nc, avg_grp_size, ICC, sqrt = TRUE)
}
\arguments{
\item{N_total}{Numerical value indicating the total sample size of the study.}

\item{Nc}{Numerical value indicating the sample size of the arm/group that does not contain clustering.}

\item{avg_grp_size}{Numerical value indicating the average cluster size.}

\item{ICC}{Numerical value indicating the intra-class correlation (ICC) value.}

\item{sqrt}{Logical indicating if the square root of gamma should be calculated. Default = \code{TRUE}}
}
\value{
Returns a numerical value for the cluster-design adjustment factor, \eqn{\gamma}.
}
\description{
This function calculates a small cluster-design adjustment factor which can be used
to adjusted effect sizes (independently of how clustering is handled) and the sampling variances from cluster-design studies
that adequately handles clustering. This factor can be found in second term of Equation (5)
from from Hedges & Citkowicz (2015, p. 1298). The factor in denoted as \eqn{\gamma}{\gamma} in WWC (2021).
The same notion is used here and gave name to the function.
}
\details{
When calculating effect sizes from cluster-designed studies, it recommended
(Hedges, 2007, 2011; Hedges & Citkowitz, 2015; WWC, 2021) to add an adjustment factor, \eqn{\gamma}{\gamma}
to \eqn{d}{d} whether or not cluster is adequately handled in the studies. Even if clustering
is adequately handled, WWC also recommend to use \eqn{\gamma}{\gamma} as a small number of clusters correction
to the variance component. The adjustment factor gamma when there is clustering in one
treatment group only is given by

\deqn{\gamma  = 1 - \dfrac{(N^C+n-2)\rho}{N_2}}

where \eqn{N^C}{N-C} is the sample size of the group without clustering, \eqn{n}{n} is
the average cluster size, and \eqn{\rho}{\rho} is the (often imputed) intraclass correlation.

Then let the naive estimator of Hedges' \eqn{g}{g} be

\deqn{g_{naive} = J\times \left(\dfrac{\bar{Y}^T_{\bullet\bullet} - \bar{Y}^C_{\bullet}}{S_T} \right)}

where \eqn{J = 1 - 3/(4df-1)}, \eqn{\bar{Y}^T_{\bullet\bullet}} it the average treatment effect for the
treatment group containing clustering, \eqn{\bar{Y}^C_{\bullet}} is the average treatment effect
for the group without clustering, and \eqn{S_T} is the standard deviation ignoring clustering. To account for the
fact that \eqn{S_T} systematically underestimates the true standard deviation, \eqn{\sigma_T},
the cluster-adjusted effect size can be obtained from

\deqn{g_T = g_{naive}\sqrt{\gamma}}

if clustering is properly adjusted for, the sampling variance of \eqn{g_T} is given by

\deqn{v_{g^T} = \left(\dfrac{1}{N^T} + \dfrac{1}{N^C}\right) \gamma + \dfrac{g^2_T}{2h} }

where \eqn{N^T} is the sample size of the treatment group containing clustering and \eqn{h} is
given by

\deqn{ h = \dfrac{[(N-2)(1-\rho) + (N^T-n)\rho]^2}
{(N-2)(1-\rho)^2 + (N^T-n)n\rho^2 + 2(N^T-n)(1-\rho)\rho}}

where \eqn{N}{N} is the total sample size.
}
\note{
Read Taylor et al. (2020) to understand why we use the \eqn{g_T}{g-T} notation.
Find suggestions for how and which ICC values to impute when these are unknown (Hedges & Hedberg, 2007, 2013).
}
\examples{

N <- 100
Nc <- 40
n <- 5
rho <- 0.1

gamma_1armcluster(N_total = N, Nc = Nc, avg_grp_size = n, ICC = rho)

# Testing function
sqrt(1 - (((Nc + n-2)*rho)/(N-2))) |> round(3)

}
\references{
Hedges, L. V. (2007).
Effect sizes in cluster-randomized designs.
\emph{Journal of Educational and Behavioral Statistics}, 32(4), 341–370.
\doi{10.3102/1076998606298043}

Hedges, L. V. (2011).
Effect sizes in three-level cluster-randomized experiments.
\emph{Journal of Educational and Behavioral Statistics}, 36(3), 346–380.
\doi{10.3102/1076998610376617}

Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}

Hedges, L. V., & Hedberg, E. C. (2007).
Intraclass correlation values for planning group-randomized trials in education.
\emph{Educational Evaluation and Policy Analysis}, 29(1), 60–87.
\doi{10.3102/0162373707299706}

Hedges, L. V., & Hedberg, E. C. (2013).
Intraclass correlations and covariate outcome correlations for planning
two- and three-Level cluster-randomized experiments in education.
\emph{Evaluation Review}, 37(6), 445–489. \doi{10.1177/0193841X14529126}

Taylor, J.A., Pigott, T.D., & Williams, R. (2020)
Promoting knowledge accumulation about intervention effects:
Exploring strategies for standardizing statistical approaches and effect size reporting.
\emph{Educational Researcher}, 51(1), 72-80. \doi{10.3102/0013189X211051319}

What Works Clearinghouse (2021).
Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1
\emph{Institute of Education Science}.
\url{https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf}
}
\seealso{
\code{\link{df_h_1armcluster}} for further details.
}
