% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_calc.R
\name{OR_calc}
\alias{OR_calc}
\title{Calculate and cluster bias adjust odds ratios (OR)}
\usage{
OR_calc(
  A = NULL,
  B = NULL,
  C = NULL,
  D = NULL,
  p1 = NULL,
  p2 = NULL,
  n1 = NULL,
  n2 = NULL,
  OR = NULL,
  LL_OR = NULL,
  UL_OR = NULL,
  SE_OR = NULL,
  V_OR = NULL,
  Z = 1.96,
  ICC = NULL,
  avg_cl_size = NULL,
  n_cluster_arms = 2,
  add_name_to_vars = NULL,
  vars = dplyr::everything()
)
}
\arguments{
\item{A}{Upper left cell of an 2 X 2 frequency table.}

\item{B}{Upper right cell of an 2 X 2 frequency table.}

\item{C}{Lower left cell of an 2 X 2 frequency table.}

\item{D}{Lower right cell of an 2 X 2 frequency table.}

\item{p1}{Risk/probability of an event in group 1 (usually the treatment group).}

\item{p2}{Risk/probability of an event in group 2 (usually the control group).}

\item{n1}{Sample size of group 1 (usually the treatment group).}

\item{n2}{Sample size of group 2 (usually the control group).}

\item{OR}{Odds ratio estimate.}

\item{LL_OR}{Lower  bound of the 95\% confidence interval of the odds ratio.}

\item{UL_OR}{Upper  bound of the 95\% confidence interval of the odds ratio.}

\item{SE_OR}{Standard error of the odds ratio.}

\item{V_OR}{Sampling variance of the odds ratio.}

\item{Z}{Z-values from an normal distribution.}

\item{ICC}{Intra-class correlation.}

\item{avg_cl_size}{Average cluster size.}

\item{n_cluster_arms}{(Optional) Number of arm with clustering.}

\item{add_name_to_vars}{Optional character string to be added to the variables names of the generated \code{tibble}.}

\item{vars}{Variables to be reported. Default is \code{NULL}. See Value section for further details.}
}
\value{
A \code{tibble} with information about OR, OR_LN, vln_OR.
}
\description{
This function calculated odds ratios based on various type of input/information,
as described in Table 11.10 from Borenstein and Hedges (2019, p. 226).
}
\details{
2x2 table\tabular{lllll}{
    \tab  \tab Event \tab Non-event \tab Total \cr
   Treated \tab  \tab A \tab B \tab n1 \cr
   Control \tab  \tab C \tab D \tab n2 \cr
}
}
\note{
Read Borenstein and Hedges (2019) for further details.
}
\examples{

# Using raw events
OR_calc(A = 20, B = 80, C = 10, D = 90)

# Using proportions
OR_calc(p1 = .2, p2 = .1, n1 = 100, n2 = 100)

# Using raw OR and CIs
OR_calc(OR = 2.25, LL_OR = 1.5, UL_OR = 3)

# Adding suffix to variables and selecting specific variables
OR_calc(A = 20, B = 80, C = 10, D = 90, add_name_to_vars = "_test", vars = OR_test)

# Cluster bias adjustment when there is clustering in both groups
OR_calc(p1 = .53, p2 = .11, n1 = 20, n2 = 26, ICC = 0.1, avg_cl_size = 8, n_cluster_arms = 2)

# Cluster bias adjustment when there is clustering in one group only
OR_calc(p1 = .53, p2 = .11, n1 = 20, n2 = 26, ICC = 0.1, avg_cl_size = 8, n_cluster_arms = 1)


}
\references{
Borenstein and Hedges (2019).
Effect sizes for meta-analysis. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.),
\emph{The handbook of research synthesis and meta-analysis} (3rd ed., pp. 207–242).
Russell Sage Foundation West Sussex.

Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}

Higgins, J. P. T., Eldridge, S., & Li, T. (2019).
In J. P. T. Higgins, J. Thomas, J. Chandler, M. S. Cumpston, T. Li, M. Page, & V. Welch (Eds.),
\emph{Cochrane handbook for systematic reviews of interventions} (2nd ed., pp. 569–593).
Wiley Online Library. \doi{10.1002/9781119536604.ch23}
}
