% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgt_smd_1armcluster.R
\name{vgt_smd_1armcluster}
\alias{vgt_smd_1armcluster}
\title{Variance calculation when there is clustering in one treatment group only}
\usage{
vgt_smd_1armcluster(
  N_cl_grp,
  N_ind_grp,
  avg_grp_size,
  ICC,
  g,
  model = c("Posttest", "ANCOVA", "emmeans", "DiD", "Reg_coef", "Std_reg_coef"),
  not_cluster_adj = TRUE,
  prepost_cor = NULL,
  F_val = NULL,
  t_val = NULL,
  SE = NULL,
  SD = NULL,
  SE_std = NULL,
  R2 = NULL,
  q = 1,
  add_name_to_vars = NULL,
  vars = dplyr::everything()
)
}
\arguments{
\item{N_cl_grp}{Numerical value indicating the sample size of the arm/group containing clustering.}

\item{N_ind_grp}{Numerical value indicating the sample size of the arm/group that does not contain clustering.}

\item{avg_grp_size}{Numerical value indicating the average cluster size.}

\item{ICC}{Numerical value indicating the (imputed) intra-class correlation.}

\item{g}{Numerical values indicating the estimated effect size (Hedges' g).}

\item{model}{Character indicating from what model the effect size estimate is
obtained. See details.}

\item{not_cluster_adj}{Logical indicating if clustering was inadequately handled in model/study. Default is \code{TRUE}.}

\item{prepost_cor}{Numerical value indicating the pre-posttest correlation.}

\item{F_val}{Numerical value indicating the reported F statistics value. Note that \eqn{F = t^2}.}

\item{t_val}{Numerical value indicating the reported t statistics value.}

\item{SE}{Numerical value indicating the (reported) non-standardized standard error.}

\item{SD}{Numerical value indicating the pooled standard deviation.}

\item{SE_std}{Numerical value indicating the (reported) standardized standard error (SE).}

\item{R2}{Numerical value indicating the (reported) \eqn{R^2} value from analysis model.}

\item{q}{Numerical value indicating the number of covariates.}

\item{add_name_to_vars}{Optional character string to be added to the variables names of the generated \code{tibble}.}

\item{vars}{Variables to be reported. Default is \code{NULL}. See Value section for further details.}
}
\value{
When \code{add_name_to_vars = NULL}, the function returns a \code{tibble} with the following variables:\if{html}{\out{<br>}}
\item{gt}{The cluster and small sample adjusted effect size estimate.}
\item{vgt}{The cluster adjusted sampling variance estimate of \eqn{gt}.}
\item{Wgt}{The cluster adjusted samplingvariance estimate of \eqn{gt}, without the second term of the variance formula, as
given in Eq. (2) in Pustejovsky & Rodgers (2019).}
\item{hg}{The variance-stabilizing transformed effect size. See Eq. (3) in Pustejovsky & Rodgers (2019)}
\item{vhg}{The approximate sampling variance of hg}
\item{h}{The degrees of freedom given in Eq (7) in Hedges & Citkowicz (2015, p. 1298). See \code{\link{df_h_1armcluster}}.}
\item{df}{The degrees of freedom. If none or one covariate \eqn{df = h}.
otherwise with two or more covariates \eqn{df = h - q}.}
\item{n_covariates}{The number of covariates in the model, defined as \eqn{q} in Hedges et al. (2023).}
\item{var_term1}{Unadjusted measure of the first term of the variance formula.}
\item{adj_fct}{Indicating whether \eqn{\eta} or \eqn{\gamma} were used to adjust the variance. That is
whether the studies handle clustering inadequately or not.}
\item{adj_value}{Estimated value of adjustment factor.}
}
\description{
This function calculates the sampling variance estimates for effect sizes
obtained from cluster-designed studies. This include measures of the sampling variance,
a modified variance estimate for publication bias testing, and a variance-stabilized
transformed effect size and variance as presented in Pustejovsky & Rodgers (2019).
}
\details{
Table 1 illustrates all cluster adjustment
of variance estimates from pre-test and/or covariate-adjusted measures that can
be calculated with the vgt_smd_1armcluster()

\emph{\strong{Table 1}}\if{html}{\out{<br>}}
\emph{Sampling variance estimates for \eqn{g_T} across various models for handling cluster, estimation techniques, and reported quantities.}\tabular{lll}{
   \strong{Calculation type/\if{html}{\out{<br>}}reported quantities} \tab \strong{Cluster-adjusted (model)\if{html}{\out{<br>}}sampling variance} \tab \strong{Not cluster-adjusted (model)\if{html}{\out{<br>}}sampling variance} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{R^2, N^T, N^C} \tab \eqn{(1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{(1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2(h-q)}.} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{R^2_{imputed}, N^T, N^C} \tab \eqn{(1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{(1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2(h-q)}.} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{F, (t^2), N^T, N^C} \tab \eqn{\left(\frac{g^2_T}{F}\right) \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{\left(\frac{g^2_T}{F}\right) \eta + \frac{g^2_T}{2(h-q)}.} \cr
   ANCOVA, pretest only\if{html}{\out{<br>}}\eqn{F, (t^2), N^T, N^C} \tab \eqn{\left(\frac{g^2_T}{F}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{\left(\frac{g^2_T}{F}\right) \eta + \frac{g^2_T}{2h}.} \cr
   ANCOVA, pretest only\if{html}{\out{<br>}}\eqn{r, N^T, N^C} \tab \eqn{(1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{(1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.} \cr
   Reg coef\if{html}{\out{<br>}}\eqn{SE, S_T, N^T, N^C} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \eta + \frac{g^2_T}{2(h-q)}.} \cr
   Reg coef, pretest only\if{html}{\out{<br>}}\eqn{SE, S_T, N^T, N^C} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \gamma + \frac{g^2_T}{2h}.} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \eta + \frac{g^2_T}{2h}.} \cr
   Std. reg coef\if{html}{\out{<br>}}\eqn{SE_{std}, N^T, N^C} \tab \eqn{SE^2_{std} \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{SE^2_{std} \eta + \frac{g^2_T}{2(h-q)}.} \cr
   Std. reg coef, pretest only\if{html}{\out{<br>}}\eqn{SE_{std}, N^T, N^C} \tab \eqn{SE^2_{std} \gamma + \frac{g^2_T}{2h}.} \tab \eqn{SE^2_{std} \eta + \frac{g^2_T}{2h}.} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{r, N^T, N^C} \tab \eqn{2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{r_{imputed}, N^T, N^C} \tab \eqn{2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{t, N^T, N^C} \tab \eqn{\left(\frac{g^2}{t^2}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{\left(\frac{g^2}{t^2}\right) \eta + \frac{g^2_T}{2h}.} \cr
}


\emph{Note}: \eqn{R^2} "is the multiple correlation between the covariates and the outcome" (WWC, 2021),
\eqn{\eta = 1 - (N^C+n-2)\rho/(N-2)}, see \code{\link{eta_1armcluster}},
\eqn{\gamma = 1 - (N^C+n-2)\rho/(N-2)}, see \code{\link{eta_1armcluster}},
\eqn{r} is the pre-posttest correlation, and \eqn{q} is the number of covariates. Std. = standardized.

"It is often desired in practice to adjust for multiple baseline characteristics.
The problem of \eqn{q} covariates is a straightforward extension of the single covariate case
(...): The correlation coefficient estimate \eqn{r} is now obtained by
taking the square root of the coefficient of multiple determination, \eqn{R^2}"
(Hedges et al. 2023, p. 17) and \eqn{df = h-q}.

\strong{Calculating modified measures of variance  for publication bias testing}

\emph{\strong{Table 2}}\if{html}{\out{<br>}}
\emph{Sampling variance estimates for \eqn{g_T} across various models for handling cluster, estimation techniques, and reported quantities.}\tabular{lll}{
   \strong{Calculation type/\if{html}{\out{<br>}}reported quantities} \tab \strong{Cluster-adjusted (model)\if{html}{\out{<br>}}sampling variance} \tab \strong{Not cluster-adjusted (model)\if{html}{\out{<br>}}sampling variance} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{R^2, N^T, N^C} \tab \eqn{(1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma} \tab \eqn{(1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta } \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{R^2_{imputed}, N^T, N^C} \tab \eqn{(1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma} \tab \eqn{(1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta } \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{F, (t^2), N^T, N^C} \tab \eqn{\left(\frac{g^2_T}{F}\right) \gamma} \tab \eqn{\left(\frac{g^2_T}{F}\right) \eta} \cr
   ANCOVA, pretest only\if{html}{\out{<br>}}\eqn{F, (t^2), N^T, N^C} \tab \eqn{\left(\frac{g^2_T}{F}\right) \gamma} \tab \eqn{\left(\frac{g^2_T}{F}\right) \eta} \cr
   ANCOVA, pretest only\if{html}{\out{<br>}}\eqn{r, N^T, N^C} \tab \eqn{(1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma} \tab \eqn{(1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta} \cr
   Reg coef\if{html}{\out{<br>}}\eqn{SE, S_T, N^T, N^C} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \gamma } \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \eta} \cr
   Reg coef, pretest only\if{html}{\out{<br>}}\eqn{SE, S_T, N^T, N^C} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \gamma} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \eta} \cr
   Std. reg coef\if{html}{\out{<br>}}\eqn{SE_{std}, N^T, N^C} \tab \eqn{SE^2_{std} \gamma } \tab \eqn{SE^2_{std} \eta} \cr
   Std. reg coef, pretest only\if{html}{\out{<br>}}\eqn{SE_{std}, N^T, N^C} \tab \eqn{SE^2_{std} \gamma} \tab \eqn{SE^2_{std} \eta} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{r, N^T, N^C} \tab \eqn{2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma} \tab \eqn{2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{r_{imputed}, N^T, N^C} \tab \eqn{2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma} \tab \eqn{2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{t, N^T, N^C} \tab \eqn{\left(\frac{g^2}{t^2}\right) \gamma} \tab \eqn{\left(\frac{g^2}{t^2}\right) \eta} \cr
}


\emph{Note}: \eqn{R^2} "is the multiple correlation between the covariates and the outcome" (WWC, 2021),
\eqn{\eta = 1 - (N^C+n-2)\rho/(N-2)}, see \code{\link{eta_1armcluster}},
\eqn{\gamma = 1 - (N^C+n-2)\rho/(N-2)}, see \code{\link{eta_1armcluster}}, and
\eqn{r} is the pre-posttest correlation. Std. = standardized.

"It is often desired in practice to adjust for multiple baseline characteristics.
The problem of \eqn{q} covariates is a straightforward extension of the single covariate case
(...): The correlation coefficient estimate \eqn{r} is now obtained by
taking the square root of the coefficient of multiple determination, \eqn{R^2}"
(Hedges et al. 2023, p. 17) and \eqn{df = h-q}.
}
\note{
Insert
}
\examples{
vgt_smd_1armcluster(
N_cl_grp = 60, N_ind_grp = 40, avg_grp_size = 10, ICC = 0.1, g = 0.2,
model = "ANCOVA", not_cluster_adj = TRUE, R2 = 0.5, q = 3
)

# Example showing how to add a suffix to the variable names
vgt_smd_1armcluster(
N_cl_grp = 60, N_ind_grp = 40, avg_grp_size = 10, ICC = 0.3, g = 0.2,
model = "ANCOVA", not_cluster_adj = TRUE, R2 = 0.5, q = 3, add_name_to_vars = "_icc_0_3"
)


}
\references{
Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}

Pustejovsky, J. E., & Rodgers, M. A. (2019).
Testing for funnel plot asymmetry of standardized mean differences.
\emph{Research Synthesis Methods}, 10(1), 57–71.
\doi{10.1002/jrsm.1332}

What Works Clearinghouse (2021).
Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1
\emph{Institute of Education Science}.
\url{https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf}
}
\seealso{
\code{\link{df_h_1armcluster}}, \code{\link{eta_1armcluster}},
\code{\link{gamma_1armcluster}}
}
