% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgt_smd_1armcluster.R
\name{vgt_smd_1armcluster}
\alias{vgt_smd_1armcluster}
\title{Variance calculation when there is clustering in one treatment group only}
\usage{
vgt_smd_1armcluster(
  N_cl_grp,
  N_ind_grp,
  avg_grp_size,
  ICC,
  g,
  model = c("Posttest", "ANCOVA", "DiD", "Reg_coef", "Std_reg_coef"),
  not_cluster_adj = TRUE,
  prepost_cor = NULL,
  F_val = NULL,
  t_val = NULL,
  SE = NULL,
  SD = NULL,
  SE_std = NULL,
  R2 = NULL,
  q = 1
)
}
\arguments{
\item{N_cl_grp}{Insert}

\item{N_ind_grp}{Insert}

\item{avg_grp_size}{Insert}

\item{ICC}{Insert}

\item{g}{Insert}

\item{model}{Insert}

\item{not_cluster_adj}{Insert}

\item{prepost_cor}{Insert}

\item{F_val}{Insert}

\item{t_val}{Insert}

\item{SE}{Insert}

\item{SD}{Insert}

\item{SE_std}{Insert}

\item{R2}{Insert}

\item{q}{Insert}
}
\value{
A \code{tibble} with info ...(Insert)...
}
\description{
Insert
}
\details{
Table 1 illustrates all cluster adjustment
of variance estimates from pre-test and/or covariate adjusted measures that can
be calculated with the vgt_smd_1armcluster()

\emph{\strong{Table 1}}\if{html}{\out{<br>}}
\emph{Sampling variance estimates for \eqn{g_T} across various models for handling cluster, estimation techniques, and reported quantities.}\tabular{lll}{
   \strong{Calculation type/\if{html}{\out{<br>}}reported quantities} \tab \strong{Cluster-adjusted (model)\if{html}{\out{<br>}}sampling variance} \tab \strong{Not cluster-adjusted (model)\if{html}{\out{<br>}}sampling variance} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{R^2, N^T, N^C} \tab \eqn{(1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{(1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2(h-q)}.} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{R^2_{imputed}, N^T, N^C} \tab \eqn{(1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{(1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2(h-q)}.} \cr
   ANCOVA, adj. means\if{html}{\out{<br>}}\eqn{F, (t^2), N^T, N^C} \tab \eqn{\left(\frac{g^2_T}{F}\right) \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{\left(\frac{g^2_T}{F}\right) \eta + \frac{g^2_T}{2(h-q)}.} \cr
   ANCOVA, pretest only\if{html}{\out{<br>}}\eqn{F, (t^2), N^T, N^C} \tab \eqn{\left(\frac{g^2_T}{F}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{\left(\frac{g^2_T}{F}\right) \eta + \frac{g^2_T}{2h}.} \cr
   ANCOVA, pretest only\if{html}{\out{<br>}}\eqn{r, N^T, N^C} \tab \eqn{(1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{(1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.} \cr
   Reg coef\if{html}{\out{<br>}}\eqn{SE, S_T, N^T, N^C} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \eta + \frac{g^2_T}{2(h-q)}.} \cr
   Reg coef, pretest only\if{html}{\out{<br>}}\eqn{SE, S_T, N^T, N^C} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \gamma + \frac{g^2_T}{2h}.} \tab \eqn{\left(\frac{SE}{S_T}\right)^2 \eta + \frac{g^2_T}{2h}.} \cr
   Std. reg coef\if{html}{\out{<br>}}\eqn{SE_{std}, N^T, N^C} \tab \eqn{SE^2_{std} \gamma + \frac{g^2_T}{2(h-q)}.} \tab \eqn{SE^2_{std} \eta + \frac{g^2_T}{2(h-q)}.} \cr
   Std. reg coef, pretest only\if{html}{\out{<br>}}\eqn{SE_{std}, N^T, N^C} \tab \eqn{SE^2_{std} \gamma + \frac{g^2_T}{2h}.} \tab \eqn{SE^2_{std} \eta + \frac{g^2_T}{2h}.} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{r, N^T, N^C} \tab \eqn{2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{r_{imputed}, N^T, N^C} \tab \eqn{2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.} \cr
   DiD, gain scores\if{html}{\out{<br>}}\eqn{t, N^T, N^C} \tab \eqn{\left(\frac{g^2}{t^2}\right) \gamma + \frac{g^2_T}{2h}.} \tab \eqn{\left(\frac{g^2}{t^2}\right) \eta + \frac{g^2_T}{2h}.} \cr
}


\emph{Note}: \eqn{R^2} "is the multiple correlation between the covariates and the outcome" (WWC, 2021),
\eqn{\eta = 1 - (N^C+n-2)\rho/(N-2)}, see \code{\link{eta_1armcluster}},
\eqn{\gamma = 1 - (N^C+n-2)\rho/(N-2)}, see \code{\link{eta_1armcluster}},
\eqn{r} is the pre-posttest correlation, and \eqn{q} is the number of covariates. Std. = standardized.

"It is often desired in practice to adjust for multiple baseline characteristics.
The problem of \eqn{q} covariates is a straightforward extension of the single covariate case
(...): The correlation coefficient estimate \eqn{r} is now obtained by
taking the square root of the coefficient of multiple determination, \eqn{R^2}"
(Hedges et al. 2023, p. 17) and \eqn{df = h-q}.
}
\note{
Insert
}
\examples{
vgt_smd_1armcluster(
N_cl_grp = 60, N_ind_grp = 40, avg_grp_size = 10, ICC = 0.1, g = 0.2,
model = "ANCOVA", not_cluster_adj = TRUE, R2 = 0.5, q = 3
)

}
\references{
Hedges, L. V., & Citkowicz, M (2015).
Estimating effect size when there is clustering in one treatment groups.
\emph{Behavior Research Methods}, 47(4), 1295-1308. \doi{10.3758/s13428-014-0538-z}

What Works Clearinghouse (2021).
Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1
\emph{Institute of Education Science}.
\url{https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf}
}
\seealso{
\code{\link{df_h_1armcluster}}, \code{\link{eta_1armcluster}},
\code{\link{gamma_1armcluster}}
}
